backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==1.4.41
psycopg2-binary==2.9.3
pydantic==1.10.2
passlib==1.7.4
python-dotenv==0.20.0
email-validator==1.2.1
